--- src/modules/transaction/components/TransactionResult/index.js
@@ -1,1 +1,0 @@
-export { default } from './TransactionResult';
