+++ src/modules/transaction/components/TxBroadcaster/illustrationsMap.js
